refactor viewer to separate component
