+++ src/App.sass
@@ -151,3 +151,12 @@ ul
     &.pending
       background-color: $denim
       color: $white
+
+#instant-coverage
+  p
+    color: $steel
+  .tickets
+    background-color: $white
+    width: 50%
+    margin: auto
+    padding: 15px
